[NIS-13910] Made connection with auth server

+++ Profiles/Profiles.Application/Services/DoctorService.cs
@@ -36,27 +36,43 @@ namespace Profiles.Application.Services
         {
             try
             {
-                _unitOfWork.BeginTransaction(cancellationToken: cancellationToken);
-                var createdAccount = await _accountService.Create(createAccountModel, cancellationToken);
-
-                var doctor = _mapper.Map<Doctor>(createDoctorModel);
-
-                doctor.AccountId = createdAccount.Id;
-
-                await _unitOfWork.DoctorRepository.CreateAsync(doctor, cancellationToken);
-
-                if (fileModel != null)
+                using (var transaction = _unitOfWork.BeginTransaction(cancellationToken: cancellationToken))
                 {
-                    await _fileService.Upload(fileModel.FileName, fileModel.FileStream, fileModel.ContentType);
-                
-                    doctor.Account.PhotoFileName = fileModel.FileName;
-                }
 
-                await _unitOfWork.SaveChangesAsync(cancellationToken);
+                    var specializationRelatedToDoctor = await _unitOfWork.SpecializationRepository.GetAsync(createDoctorModel.SpecializationId, cancellationToken: cancellationToken);
+                    if (specializationRelatedToDoctor == null || !specializationRelatedToDoctor.IsActive)
+                    {
+                        throw new RelatedObjectNotFoundException($"Related specialization with id {createDoctorModel.SpecializationId} is not found or not active.");
+                    }
+
+                    var officeRelatedToDoctor = await _unitOfWork.OfficeRepository.GetAsync(createDoctorModel.OfficeId, cancellationToken: cancellationToken);
+                    if (officeRelatedToDoctor == null || !officeRelatedToDoctor.IsActive)
+                    {
+                        throw new RelatedObjectNotFoundException($"Related office with id {createDoctorModel.OfficeId} is not found or not active.");
+                    }
+
+                    var createdAccount = await _accountService.Create(createAccountModel, cancellationToken);
+
+                    var doctor = _mapper.Map<Doctor>(createDoctorModel);
+
+                    doctor.AccountId = createdAccount.Id;
+
+                    await _unitOfWork.DoctorRepository.CreateAsync(doctor, cancellationToken);
+
+                    if (fileModel != null)
+                    {
+                        await _fileService.Upload(fileModel.FileName, fileModel.FileStream, fileModel.ContentType);
+
+                        doctor.Account.PhotoFileName = fileModel.FileName;
+                    }
+
+                    await transaction.CommitAsync(cancellationToken);
+                    await _unitOfWork.SaveChangesAsync(cancellationToken);
+                }
             }
             catch
             {
-                if (fileModel != null)
+                if (fileModel != null && await _fileService.DoesFileExist(fileModel.FileName))
                 {                
                     await _fileService.Remove(fileModel.FileName);
                 }
@@ -67,7 +83,7 @@ namespace Profiles.Application.Services
 
         public async Task Delete(Guid id, CancellationToken cancellationToken)
         {
-            var doctorToDelete = await _unitOfWork.DoctorRepository.GetAsync(id, cancellationToken);
+            var doctorToDelete = await _unitOfWork.DoctorRepository.GetAsync(id, cancellationToken: cancellationToken);
             if (doctorToDelete == null)
             {
                 throw new NotFoundException($"Doctor with id: {id} is not found. Can't delete.");
@@ -77,17 +93,17 @@ namespace Profiles.Application.Services
 
             await _fileService.Remove(doctorToDelete.Account.PhotoFileName);
 
-            await _unitOfWork.SaveChangesAsync(cancellationToken);
+            await _unitOfWork.SaveChangesAsync(cancellationToken: cancellationToken);
         }
 
         public async Task<DoctorModel> Get(Guid id, CancellationToken cancellationToken)
         {
-            return _mapper.Map<DoctorModel>(await _unitOfWork.DoctorRepository.GetAsync(id, cancellationToken));
+            return _mapper.Map<DoctorModel>(await _unitOfWork.DoctorRepository.GetAsync(id, cancellationToken: cancellationToken));
         }
 
         public async Task<IEnumerable<DoctorModel>> GetAll(CancellationToken cancellationToken)
         {
-            return _mapper.Map<IEnumerable<DoctorModel>>(await _unitOfWork.DoctorRepository.GetAllAsync(cancellationToken));
+            return _mapper.Map<IEnumerable<DoctorModel>>(await _unitOfWork.DoctorRepository.GetAllAsync(cancellationToken: cancellationToken));
         }
 
         public async Task Update(
@@ -95,15 +111,36 @@ namespace Profiles.Application.Services
             FileModel? fileModel, 
             CancellationToken cancellationToken)
         {
-            var doctorToUpdate = await _unitOfWork.DoctorRepository.GetAsync(updateDoctorModel.Id, cancellationToken)
+            var specializationRelatedToDoctor = await _unitOfWork.SpecializationRepository.GetAsync(updateDoctorModel.SpecializationId, cancellationToken: cancellationToken);
+            if (specializationRelatedToDoctor == null || !specializationRelatedToDoctor.IsActive)
+            {
+                throw new RelatedObjectNotFoundException($"Related specialization with id {updateDoctorModel.SpecializationId} is not found or not active.");
+            }
+
+            var officeRelatedToDoctor = await _unitOfWork.OfficeRepository.GetAsync(updateDoctorModel.OfficeId, cancellationToken: cancellationToken);
+            if (officeRelatedToDoctor == null || !officeRelatedToDoctor.IsActive)
+            {
+                throw new RelatedObjectNotFoundException($"Related office with id {updateDoctorModel.OfficeId} is not found or not active.");
+            }
+
+            var doctorToUpdate = await _unitOfWork.DoctorRepository.GetAsync(updateDoctorModel.Id, isIncluded: false, cancellationToken: cancellationToken)
                 ?? throw new NotFoundException($"Doctor with id: {updateDoctorModel.Id} is not found. Can't update.");
 
             if (fileModel != null)
             {
-                await _fileService.Remove(doctorToUpdate.Account.PhotoFileName);
+                var accountToUpdate = await _unitOfWork.AccountRepository.GetAsync(doctorToUpdate.AccountId, cancellationToken: cancellationToken);
+
+                if (accountToUpdate == null)
+                {
+                    throw new RelatedObjectNotFoundException($"Related account with id {doctorToUpdate.AccountId} is not found.");
+                }
+
+                await _fileService.Remove(accountToUpdate.PhotoFileName);
                 await _fileService.Upload(fileModel.FileName, fileModel.FileStream, fileModel.ContentType);
-                
-                doctorToUpdate.Account.PhotoFileName = fileModel.FileName;
+
+                accountToUpdate.PhotoFileName = fileModel.FileName;
+
+                await _unitOfWork.AccountRepository.UpdateAsync(accountToUpdate, updateDoctorModel.AuthorId, cancellationToken);
             }
 
             _mapper.Map(updateDoctorModel, doctorToUpdate);
